utils -> libraries rename and GUI script moved to postprocessing component
