+++ VAREID/algo/detection/video_detector.py
@@ -13,7 +13,7 @@ import ultralytics
 import yaml
 from tqdm import tqdm
 
-from VAREID.util.utils import path_from_file
+from VAREID.libraries.utils import path_from_file
 
 ultralytics.checks()
 import urllib.request
@@ -22,7 +22,7 @@ from pathlib import Path
 import pandas as pd
 from ultralytics import YOLO
 
-from VAREID.util.io.format_funcs import clone_from_github, load_config, load_json, save_json, split_dataframe
+from VAREID.libraries.io.format_funcs import clone_from_github, load_config, load_json, save_json, split_dataframe
 
 warnings.filterwarnings("ignore")
