+++ VAREID/algo/frame_sampling/frame_sampling.py
@@ -7,8 +7,8 @@ from copy import deepcopy
 import pandas as pd
 import shutil
 
-from VAREID.util.io.format_funcs import load_config, load_json, save_json, split_dataframe, join_dataframe_dict
-from VAREID.util.utils import path_from_file
+from VAREID.libraries.io.format_funcs import load_config, load_json, save_json, split_dataframe, join_dataframe_dict
+from VAREID.libraries.utils import path_from_file
 
 def group_annotations_by_tracking_id_and_subsequences(data):
     annotations = data['annotations']
