+++ VAREID/algo/ia_classification/IA_classifier.py
@@ -17,8 +17,8 @@ from torchvision.transforms import functional as F
 from torchvision.models import resnet50
 from torchvision.ops import nms
 
-from VAREID.util.io.format_funcs import load_config, load_json, save_json, split_dataframe, join_dataframe
-from VAREID.util.utils import path_from_file
+from VAREID.libraries.io.format_funcs import load_config, load_json, save_json, split_dataframe, join_dataframe
+from VAREID.libraries.utils import path_from_file
 
 
 def xywh_to_xyxy(bbox: list):
