+++ VAREID/algo/viewpoint_classification/viewpoint_classifier.py
@@ -14,8 +14,8 @@ from albumentations import Compose, Normalize, Resize
 from albumentations.pytorch import ToTensorV2
 from torch.utils.data import Dataset
 
-from VAREID.util.io.format_funcs import load_config, load_json, save_json, split_dataframe, join_dataframe
-from VAREID.util.utils import path_from_file
+from VAREID.libraries.io.format_funcs import load_config, load_json, save_json, split_dataframe, join_dataframe
+from VAREID.libraries.utils import path_from_file
 
 # Load configuration
 config = load_config(path_from_file(__file__, "viewpoint_classifier_config.yaml"))
