+++ VAREID/drivers/import_image_driver.py
@@ -1,8 +1,8 @@
 import argparse
 
-from VAREID.util.io.format_funcs import load_config
-from VAREID.util.io.logging import setup_logging, log_subprocess
-from VAREID.util.io.workflow_funcs import build_config, decode_config
+from VAREID.libraries.io.format_funcs import load_config
+from VAREID.libraries.io.logging import setup_logging, log_subprocess
+from VAREID.libraries.io.workflow_funcs import build_config, decode_config
 
 def main(args):
     # SELECT THE CORRECT CONFIG
