+++ VAREID/drivers/post_driver.py
@@ -3,9 +3,9 @@ from multiprocessing import Process
 import os
 from pathlib import Path
 
-from VAREID.util.io.format_funcs import load_config
-from VAREID.util.io.logging import log_subprocess, setup_logging
-from VAREID.util.io.workflow_funcs import build_config
+from VAREID.libraries.io.format_funcs import load_config
+from VAREID.libraries.io.logging import log_subprocess, setup_logging
+from VAREID.libraries.io.workflow_funcs import build_config
 
 def main(args):
     config = build_config(load_config(args.config_path))
@@ -18,7 +18,7 @@ def main(args):
         interaction_mode = "database"
 
     post_command = f'python -u -m VAREID.algo.postprocessing.postprocessing {config["image_dir"]} {config["post_left_in_path"]} {config["post_right_in_path"]} {config["post_left_out_path"]} {config["post_right_out_path"]} --db {config["post_db_path"]} --interaction_mode {interaction_mode}'
-    gui_command = f'python -u -m VAREID.util.ui.gui --db {config["post_db_path"]} --allowed_dir {config["data_dir_out"]}'
+    gui_command = f'python -u -m VAREID.algo.postprocessing.gui --db {config["post_db_path"]} --allowed_dir {config["data_dir_out"]}'
 
     # MODE (UI)
     if interaction_mode == "database":
