+++ VAREID/libraries/constants.py
