+++ VAREID/libraries/db/directory.py
