+++ VAREID/libraries/ggr_funcs.py
