--- config_refactor.yaml
@@ -1,94 +1,0 @@
-# CONFIGURATION FILE: PLEASE READ OUR DOCUMENTATION FOR MORE INFORMATION ON HOW THIS WORKS
-# You DO NOT need to change anything other than what is marked. Read up on the other values 
-# if you want to customize the output formatting of the pipeline.
-
-
-# -------- FILL OUT THIS SECTION --------
-
-# The file path (dir) to save your output to
-data_dir_out: /fs/ess/PAS2136/ggr_data/results/Grevys_Experiment_4_refactored
-# The file path (dir) to read data from
-data_dir_in: /fs/ess/PAS2136/ggr_data/video_data/Grevys_Experiment_4
-# True/False describing whether the data is in video format (True) or image format (False)
-data_video: True
-
-# -------- END FILL OUT SECTION --------
-
-
-# Important directories to reference
-log_dirname: "logs"
-image_dirname: "images"
-model_dirname: "VAREID/models"
-
-# Import step fields
-image_out_file: "image_data.json"
-video_out_file: "video_data.json"
-import_logfile: "import.txt"
-
-# Detection step fields
-dt_dirname: "detector"
-dt_model: "yolov10l"
-dt_logfile: "detector.txt"
-
-# Detection (gt present) fields -- FILL IN IF PROVIDING GT FILE
-dt_gt_file: ~
-dt_filtered_out_file: ~
-
-# Detection split based on image or video (snakemake needs this split)
-dt_video_out_file: "vid_annots.json"
-dt_image_out_file: "img_annots.json"
-
-# Species identification step fields
-si_dirname: "species_identifier"
-si_out_file: "si_annots.json"
-si_logfile: "species_identifier.txt"
-
-# Viewpoint classification step fields
-vc_dirname: "viewpoint_classifier"
-vc_model: "plains_grevys_VC_june_25_tf_efficientnet_b4_ns_9.pth"
-vc_out_file: "vc_annots.json"
-vc_logfile: "viewpoint_classifier.txt"
-
-# Identifiable annotation classification step fields
-ia_dirname: "ia_classifier"
-ia_model: "best_june_24_2025_IA_classifier_016.pth"
-ia_out_file: "ia_annots.json"
-ia_filtered_out_file: "ia_filtered_annots.json"
-ia_logfile: "ia_classify_and_filter.txt"
-
-# Frame sampling step fields
-fs_dirname: "frame_sampling"
-fs_out_file: "fs_annots.json"
-fs_logfile: "frame_sampling.txt"
-
-# Optional stage1 output of frame sampling -- OPTIONAL FILL IN
-fs_stage1_out_file: "stage1_fs_annots.json"
-
-# Miew-id step fields
-mid_dirname: "miew_id"
-mid_model: "conservationxlabs/miewid-msv3"
-mid_out_file: "miewid_embeddings.pickle"
-mid_logfile: "miew_id.txt"
-
-# LCA step fields
-lca_dirname: "lca"
-lca_verifiers_probs: "verifiers_probs.json"
-lca_subunit_logfile: "lca_algorithm.txt"
-lca_logfile: "lca.txt"
-
-lca_separate_viewpoints: True
-
-# LCA is special: its output is formatted as {prefix}_{viewpoint}_{suffix}.json (or {prefix}_{suffix}.json if not separating viewpoint).
-lca_out_prefix: "lca"
-lca_out_suffix: "annots"
-
-# Postprocessing step fields
-post_dirname: "post"
-post_db_file: "post_ui_decision_db.db"
-post_left_out_file: "post_left_annots.json"
-post_right_out_file: "post_right_annots.json"
-post_logfile: "post.txt"
-gui_logfile: "gui.txt"
-
-# Optional interaction mode for postprocessing human decision making. OPTIONS: [console, ipywidgets, console]
-interaction_mode: "database"
